requirements.txt
spotipy>=2.24.0
tidalapi>=0.8.3
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
colorama>=0.4.6
tqdm>=4.66.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0